requirements.txt
flask
pymupdf
opencv-python
numpy
pillow
rembg
onnxruntime
google-genai
huggingface_hub
python-dotenv
